backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0

openai==1.10.0
faiss-cpu==1.7.4
langchain==0.1.4
langchain-community==0.0.16

PyPDF2==3.0.1
python-docx==1.1.0

python-telegram-bot==20.7
slack-sdk==3.26.2

python-dotenv==1.0.0
aiofiles==23.2.1
